N21-1494 introduce composable
