+++ src/components/data-external-tool/ExternalToolLaunchState.composable.ts
@@ -1,5 +1,4 @@
 import {
-	ToolConfigurationStatus,
 	ToolLaunchRequest,
 	ToolLaunchRequestMethodEnum,
 } from "@/store/external-tool";
@@ -91,25 +90,6 @@ export const useExternalToolLaunchState = () => {
 		document.body.appendChild(form);
 		form.submit();
 	};
-	const determineOutdatedTranslationKey = (
-		userRoles: string[],
-		toolConfigStatus: ToolConfigurationStatus | undefined
-	): string => {
-		if (userRoles.includes("teacher")) {
-			if (
-				toolConfigStatus?.isOutdatedOnScopeContext &&
-				toolConfigStatus.isOutdatedOnScopeSchool
-			) {
-				return "common.information.outdatedOnSchoolAndContext.teacher";
-			} else if (toolConfigStatus?.isOutdatedOnScopeSchool) {
-				return "common.information.outdatedOnSchool.teacher";
-			} else {
-				return "common.information.outdatedOnContext.teacher";
-			}
-		} else {
-			return "common.information.outdated.student";
-		}
-	};
 
 	return {
 		toolLaunchRequest,
@@ -117,6 +97,5 @@ export const useExternalToolLaunchState = () => {
 		isLoading,
 		fetchLaunchRequest,
 		launchTool,
-		determineOutdatedTranslationKey,
 	};
 };
