+++ src/components/data-external-tool/index.ts
@@ -3,3 +3,4 @@ export * from "./ExternalToolElementDisplayState.composable";
 export * from "./ExternalToolLaunchState.composable";
 export * from "./SchoolExternalToolApi.composable";
 export * from "./SchoolExternalToolUsage.composable";
+export * from "./ToolConfigurationStatus.composable";
