+++ src/locales/uk.json
@@ -1068,7 +1068,6 @@
   "feature-board-external-tool-element.dialog.title": "Вибір і налаштування",
   "feature-board-external-tool-element.alert.error.teacher": "Інструмент зараз неможливо запустити. Будь ласка, оновіть дошку або зверніться до адміністратора школи.",
   "feature-board-external-tool-element.alert.error.student": "Інструмент зараз неможливо запустити. Будь ласка, оновіть дошку або зверніться до вчителя чи інструктора курсу.",
-  "feature-board-external-tool-element.alert.outdated.teacher": "Конфігурація інструменту застаріла, тому інструмент не можна запустити. Щоб оновити, зверніться до адміністратора школи.",
   "common.information.outdatedOnContext.teacher": "Конфігурація інструменту застаріла, тому інструмент не можна запустити. Оновіть конфігурацію інструмента.",
   "common.information.outdatedOnSchool.teacher": "Конфігурація інструменту застаріла, тому інструмент не можна запустити. Для оновлення зверніться до адміністратора школи.",
   "common.information.outdatedOnSchoolAndContext.teacher": "Конфігурація інструменту застаріла, тому інструмент не можна запустити. Оновіть конфігурацію інструменту та зверніться до адміністратора школи.",
